+++ packages/studio-explore/src/components/Statistics/Lables/index.tsx
@@ -1,52 +1,17 @@
 import React, { useEffect, useState } from 'react';
-import Chart from '../../ChartView/index';
-import { useContext, IQueryStatement } from '@graphscope/studio-graph';
-import { Flex, Card } from 'antd';
+
+import { useContext, IQueryStatement, SchemaView, type GraphData } from '@graphscope/studio-graph';
+
+import { Flex, Typography } from 'antd';
 interface ILabelsProps {}
 
 const Labels: React.FunctionComponent<ILabelsProps> = props => {
-  const { store } = useContext();
-  const { schema, getService } = store;
-  const [state, setState] = useState({
-    nodes: [],
-  });
-  const query = async () => {
-    const { nodes } = schema;
-    if (nodes.length === 0) {
-      return;
-    }
-    const allFetch = nodes.map(item => {
-      const { label } = item;
-      const script = `MATCH (n:${label}) RETURN count(n) as ${label}`;
-      return getService<IQueryStatement>('queryStatement')(script);
-    });
-    Promise.all(allFetch).then(res => {
-      console.log(res);
-      const data = res.map(item => {
-        const { table } = item;
-        const [name, value] = Object.entries(table[0])[0];
-        return {
-          name,
-          value,
-        };
-      });
-      //@ts-ignore
-      setState(preState => {
-        return {
-          ...preState,
-          nodes: data,
-        };
-      });
-    });
-  };
-  useEffect(() => {
-    query();
-  }, [schema]);
-  console.log(state.nodes);
-
   return (
-    <Flex gap={12} style={{ width: '300px' }}>
-      <Chart data={state.nodes} type="pie" xField="name" yField="value" style={{ height: 120, padding: 10 }} />
+    <Flex gap={12} style={{ width: '100%', height: '200px' }} vertical>
+      <Typography.Title level={5} style={{ margin: '0px' }}>
+        Labels Statistics
+      </Typography.Title>
+      <SchemaView id="explore-schema-view" />
     </Flex>
   );
 };