+++ examples/graphy/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @graphscope/graphy-website
 
+## 0.1.8
+
+### Patch Changes
+
+- update studio graph
+- Updated dependencies
+  - @graphscope/studio-graph@0.1.14
+  - @graphscope/studio-components@0.1.14
+  - @graphscope/studio-driver@0.1.12
+  - @graphscope/studio-importor@0.1.14
+  - @graphscope/studio-query@0.1.14
+
 ## 0.1.7
 
 ### Patch Changes