+++ packages/studio-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @graphscope/studio-components
 
+## 0.1.14
+
+### Patch Changes
+
+- update studio graph
+
 ## 0.1.13
 
 ### Patch Changes